% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshuffle.R
\name{ggshuffle}
\alias{ggshuffle}
\title{ggshuffle}
\usage{
ggshuffle(plot)
}
\arguments{
\item{plot}{A ggplot2 plot object.}
}
\value{
Nothing.
}
\description{
\code{ggshuffle} is a RStudio-Addin that delivers a graphical interface for
detecting if you charts appearance change with plotting order..
}
\details{
To run the addin, either highlight a ggplot2-object in your current
    script and select \code{ggshuffle} from the Addins-menu within RStudio,
    or run \code{ggshuffle(plot)} with a ggplot2 object as the parameter.
}
\examples{
\dontrun{
# example for ggshuffle command line usage.
library(ggplot2)
gg <- ggplot(na.omit(txhousing), aes(volume, median, color = year)) +
  geom_point()

ggshuffle(ddd)
}
}
